# api/openapi.yaml
openapi: 3.1.0
info:
  title: SecureLedger API
  version: 0.1.1
  description: |
    Local reference API for security-focused ledger engineering.
    X-Principal-* headers are a development-only identity boundary and must not
    be used for an internet deployment.
  license:
    name: MIT
    identifier: MIT
servers:
  - url: http://localhost:8080
tags:
  - name: System
    description: Process health and operational endpoints.
  - name: Accounts
    description: Account creation and authorised balance reads.
  - name: Transfers
    description: Idempotent money movement.
  - name: Oversight
    description: Read-only journal, audit and risk views.
security:
  - DevelopmentPrincipalId: []
    DevelopmentPrincipalRole: []
paths:
  /healthz:
    get:
      summary: Check process health
      tags: [System]
      operationId: health
      security: []
      responses:
        "200":
          description: Process is serving requests.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Health"
  /v1/accounts:
    post:
      summary: Create an account
      tags: [Accounts]
      operationId: createAccount
      parameters:
        - $ref: "#/components/parameters/PrincipalId"
        - $ref: "#/components/parameters/PrincipalRole"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAccountRequest"
      responses:
        "201":
          description: Account created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
        "400": { $ref: "#/components/responses/BadRequest" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "413": { $ref: "#/components/responses/RequestTooLarge" }
        "415": { $ref: "#/components/responses/UnsupportedMediaType" }
  /v1/accounts/{accountId}:
    get:
      summary: Get an authorised account view
      tags: [Accounts]
      operationId: getAccount
      parameters:
        - name: accountId
          in: path
          required: true
          schema: { type: string, minLength: 1 }
        - $ref: "#/components/parameters/PrincipalId"
        - $ref: "#/components/parameters/PrincipalRole"
      responses:
        "200":
          description: Account visible to the principal.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /v1/transfers:
    post:
      summary: Commit or replay a transfer
      tags: [Transfers]
      operationId: createTransfer
      parameters:
        - $ref: "#/components/parameters/PrincipalId"
        - $ref: "#/components/parameters/PrincipalRole"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTransferRequest"
      responses:
        "201":
          description: Transfer committed for the first time.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransferResult"
        "200":
          description: Original result returned for an idempotent replay.
          headers:
            Idempotent-Replayed:
              schema: { type: string, const: "true" }
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransferResult"
        "400": { $ref: "#/components/responses/BadRequest" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409":
          description: Insufficient funds, currency mismatch, or idempotency conflict.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "413": { $ref: "#/components/responses/RequestTooLarge" }
        "415": { $ref: "#/components/responses/UnsupportedMediaType" }
        "422":
          description: Configured transfer limit exceeded.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
  /v1/journal:
    get:
      summary: List journal postings
      tags: [Oversight]
      operationId: listJournal
      parameters:
        - $ref: "#/components/parameters/PrincipalId"
        - $ref: "#/components/parameters/PrincipalRole"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: Most recent journal postings in sequence order.
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items: { $ref: "#/components/schemas/Posting" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /v1/audit:
    get:
      summary: List audit records
      tags: [Oversight]
      operationId: listAudit
      parameters:
        - $ref: "#/components/parameters/PrincipalId"
        - $ref: "#/components/parameters/PrincipalRole"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: Most recent audit records.
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items: { $ref: "#/components/schemas/AuditRecord" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /v1/risk-events:
    get:
      summary: List risk events
      tags: [Oversight]
      operationId: listRiskEvents
      parameters:
        - $ref: "#/components/parameters/PrincipalId"
        - $ref: "#/components/parameters/PrincipalRole"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: Most recent risk events.
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items: { $ref: "#/components/schemas/RiskEvent" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "403": { $ref: "#/components/responses/Forbidden" }
components:
  securitySchemes:
    DevelopmentPrincipalId:
      type: apiKey
      in: header
      name: X-Principal-ID
      description: Development-only asserted identity; not real authentication.
    DevelopmentPrincipalRole:
      type: apiKey
      in: header
      name: X-Principal-Role
      description: Development-only asserted role; not real authentication.
  parameters:
    PrincipalId:
      name: X-Principal-ID
      in: header
      required: true
      description: Development-only asserted principal identifier.
      schema: { type: string, minLength: 1 }
    PrincipalRole:
      name: X-Principal-Role
      in: header
      required: true
      description: Development-only asserted role.
      schema:
        type: string
        enum: [customer, operator, admin, auditor]
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: Scoped to the asserted principal and bound to one transfer intent.
      schema: { type: string, minLength: 8, maxLength: 128 }
    Limit:
      name: limit
      in: query
      required: false
      schema: { type: integer, minimum: 1, maximum: 1000, default: 100 }
  responses:
    BadRequest:
      description: Request syntax or values are invalid.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    Forbidden:
      description: Principal is missing, invalid, or not authorised.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    NotFound:
      description: Resource does not exist or is not externally visible.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    RequestTooLarge:
      description: Request body exceeds 1 MiB.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    UnsupportedMediaType:
      description: Content-Type is not application/json.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
  schemas:
    Health:
      type: object
      additionalProperties: false
      required: [status]
      properties:
        status: { type: string, const: ok }
    CreateAccountRequest:
      type: object
      additionalProperties: false
      required: [owner_id, currency]
      properties:
        owner_id: { type: string, minLength: 1, maxLength: 128 }
        currency: { type: string, pattern: "^[A-Z]{3}$" }
        opening_balance_minor: { type: integer, minimum: 0, default: 0 }
    Account:
      type: object
      additionalProperties: false
      required: [id, owner_id, currency, balance_minor, system, created_at]
      properties:
        id: { type: string }
        owner_id: { type: string }
        currency: { type: string, pattern: "^[A-Z]{3}$" }
        balance_minor: { type: integer, format: int64 }
        system: { type: boolean, const: false }
        created_at: { type: string, format: date-time }
    CreateTransferRequest:
      type: object
      additionalProperties: false
      required: [from_account_id, to_account_id, amount_minor]
      properties:
        from_account_id: { type: string, minLength: 1 }
        to_account_id: { type: string, minLength: 1 }
        amount_minor: { type: integer, format: int64, minimum: 1 }
        description: { type: string, maxLength: 200, default: "" }
    TransferIntent:
      type: object
      additionalProperties: false
      required: [from_account_id, to_account_id, amount_minor, description]
      properties:
        from_account_id: { type: string }
        to_account_id: { type: string }
        amount_minor: { type: integer, format: int64 }
        description: { type: string }
    Transfer:
      type: object
      additionalProperties: false
      required: [id, idempotency_key, intent, currency, actor_id, created_at]
      properties:
        id: { type: string }
        idempotency_key: { type: string }
        intent: { $ref: "#/components/schemas/TransferIntent" }
        currency: { type: string, pattern: "^[A-Z]{3}$" }
        actor_id: { type: string }
        created_at: { type: string, format: date-time }
    TransferResult:
      type: object
      additionalProperties: false
      required: [transfer, replayed]
      properties:
        transfer: { $ref: "#/components/schemas/Transfer" }
        replayed: { type: boolean }
    Posting:
      type: object
      additionalProperties: false
      required: [id, transaction_id, account_id, amount_minor, currency, sequence, created_at]
      properties:
        id: { type: string }
        transaction_id: { type: string }
        account_id: { type: string }
        amount_minor: { type: integer, format: int64 }
        currency: { type: string, pattern: "^[A-Z]{3}$" }
        sequence: { type: integer, format: int64, minimum: 1 }
        created_at: { type: string, format: date-time }
    AuditRecord:
      type: object
      additionalProperties: false
      required: [id, actor_id, action, resource_id, outcome, created_at]
      properties:
        id: { type: string }
        actor_id: { type: string }
        action: { type: string }
        resource_id: { type: string }
        outcome: { type: string }
        metadata: { type: object, additionalProperties: true }
        created_at: { type: string, format: date-time }
    RiskEvent:
      type: object
      additionalProperties: false
      required: [id, type, severity, transfer_id, reason, created_at]
      properties:
        id: { type: string }
        type: { type: string }
        severity: { type: string, enum: [low, medium, high, critical] }
        transfer_id: { type: string }
        reason: { type: string }
        created_at: { type: string, format: date-time }
    ErrorResponse:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          type: object
          additionalProperties: false
          required: [code, message]
          properties:
            code: { type: string }
            message: { type: string }